requirements.txt
dj_database_url
django-environ
gunicorn
psycopg2
pillow
